refactor: collection-related pages and components have been moved into collections module
