+++ src/front/modules/collections/collections-pages/CollectionView.jsx
