+++ src/front/modules/collections/collections-subcomponents/CollectionCard.jsx
