+++ src/front/modules/collections/collections-subcomponents/CollectionModal.jsx
